Unblock tests by skipping if concurrency is not supported Resolves rdar://101582542
