+++ Tests/SPMBuildCoreTests/PluginInvocationTests.swift
@@ -585,6 +585,9 @@ class PluginInvocationTests: XCTestCase {
     }
 
     func testPrebuildPluginShouldNotUseExecTarget() throws {
+        // Only run the test if the environment in which we're running actually supports Swift concurrency (which the plugin APIs require).
+        try XCTSkipIf(!UserToolchain.default.supportsSwiftConcurrency(), "skipping because test environment doesn't support concurrency")
+
         try testWithTemporaryDirectory { tmpPath in
             // Create a sample package with a library target and a plugin.
             let packageDir = tmpPath.appending(components: "mypkg")
@@ -718,71 +721,6 @@ class PluginInvocationTests: XCTestCase {
                     result.check(diagnostic: .contains(msg), severity: .error)
                 }
             }
-        }
-    }
-
-    func testShouldNotRequireNonPluginTarget() throws {
-        try testWithTemporaryDirectory { tmpPath in
-            // Create a sample package with a library target and a plugin.
-            let packageDir = tmpPath.appending(components: "MyPackage")
-            try localFileSystem.createDirectory(packageDir, recursive: true)
-            try localFileSystem.writeFileContents(packageDir.appending(component: "Package.swift"), string: """
-                   // swift-tools-version: 5.7
-                   import PackageDescription
-                   let package = Package(
-                       name: "MyPackage",
-                       products: [
-                           .plugin(name: "Foo", targets: ["Foo"])
-                       ],
-                       dependencies: [
-                       ],
-                       targets: [
-                           .plugin(
-                               name: "Foo",
-                               capability: .command(
-                                   intent: .custom(verb: "Foo", description: "Plugin example"),
-                                   permissions: []
-                               )
-                           )
-                       ]
-                   )
-                   """)
-
-            let myPluginTargetDir = packageDir.appending(components: "Plugins", "Foo")
-            try localFileSystem.createDirectory(myPluginTargetDir, recursive: true)
-            try localFileSystem.writeFileContents(myPluginTargetDir.appending(component: "plugin.swift"), string: """
-                     import PackagePlugin
-                     @main struct FooPlugin: BuildToolPlugin {
-                         func createBuildCommands(
-                             context: PluginContext,
-                             target: Target
-                         ) throws -> [Command] { }
-                     }
-                     """)
-
-            // Load a workspace from the package.
-            let observability = ObservabilitySystem.makeForTesting()
-            let workspace = try Workspace(
-                fileSystem: localFileSystem,
-                forRootPackage: packageDir,
-                customManifestLoader: ManifestLoader(toolchain: UserToolchain.default),
-                delegate: MockWorkspaceDelegate()
-            )
-
-            // Load the root manifest.
-            let rootInput = PackageGraphRootInput(packages: [packageDir], dependencies: [])
-            let rootManifests = try tsc_await {
-                workspace.loadRootManifests(
-                    packages: rootInput.packages,
-                    observabilityScope: observability.topScope,
-                    completion: $0
-                )
-            }
-            XCTAssert(rootManifests.count == 1, "\(rootManifests)")
-
-            // Load the package graph.
-            let _ = try workspace.loadPackageGraph(rootInput: rootInput, observabilityScope: observability.topScope)
-            XCTAssertNoDiagnostics(observability.diagnostics)
         }
     }
 }
